fix types of variables

+++ src/main/java/nflx/rozhnov/accountservice/dto/response/AccountAddBalanceRs.java
@@ -3,6 +3,7 @@ package nflx.rozhnov.accountservice.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -10,6 +11,6 @@ import java.util.UUID;
 @AllArgsConstructor
 public class AccountAddBalanceRs {
     private UUID transactionId;
-    private double newBalance;
+    private BigDecimal newBalance;
     private Date timestamp;
 }
